refactor: change name of test folder to quality_checks
