+++ nodes/web/quality_checks/check_node_group_presence.js
