+++ nodes/web/quality_checks/nodes_with_randomized_control_after_generate.js
